requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
mysql-connector-python==8.2.0
python-socketio==5.10.0
python-engineio==4.8.0
